README: Update repo name; note on Zsh; minor reformatting
This was moved on GitHub from `0cjs/script-compat-utils` to
`cynic-net/shell-compat-utils`. The renaming is because "script" was a
little vague; this is actually about shell scripts intended to run on
Bash 5.x, Bash 3.x, Zsh and perhaps even POSIX Bourne shell, and the
common utilities that they use (e.g., sed) that may differ between
systems.