Test: Automatically update all submodules; tweak bash3 build
We now get the full list of paths to submodules from .gitmodules and,
if any of those directories doesn't exist, we init the modules. This
could conceivably fail if the directories exist without the submodules
ever having being init'd, but that seems like any unlikely case. (It
wont' happen with a standard `git clone`.)

This means we need to make a slight tweak to how we decide in the bash
build to touch parse.y to regenerate the parser. We now check for a
`config.h`; if not present we assume that configure has never been run
and thus touch parse.y to force the regeneration. (We tried just
always regenerating, but that perceptably slows subsequent runs where
nothing actually needs to be rebuilt.)