Test: Check submodule status instead of just looking for empty dirs
It turns out that we _do_ end up with empty submodule directories when
we `git clone`, so we can't rely on that to tell us if the submodule
has been initialised or not. Instead, use `git submodule status`,
which prints a `-` at the start of the line for an initialised
submodule, and use that to detect whether we need to init the submodule.

This was not working well when we checked for _any_ submodules that
weren't initialised and tried a `git submodule update --init` if any
were not, so instead we do it submodule by submodule. We lose the
parallelism this way, unfortunately, but this happens only once per
clone, so it's not that big a deal.